requirements.txt
fastapi
uvicorn
pymongo
passlib[bcrypt]
python-multipart
python-jose[cryptography]